ADD: warning fix [MOLFIT-31]

+++ PdfReportKit/PRKGenerator.m
@@ -75,7 +75,7 @@ static NSArray * reportDefaultTags = nil;
         UIGraphicsBeginPDFContextToData(currentReportData, CGRectMake(0, 0, 1000, 800), nil);
         
     currentReportItemsPerPage = itemsPerPage;
-    int pagesCount = totalItems / itemsPerPage;
+    int pagesCount = (int) (totalItems / itemsPerPage);
     for (int i = 0; i < pagesCount; i++)
     {
         [renderedTags removeAllObjects];
@@ -145,7 +145,7 @@ static NSArray * reportDefaultTags = nil;
         }
         else if ([tagName isEqualToString:@"pageNumber"])
         {
-            renderedTag = [NSString stringWithFormat:@"%d", pageNumber];
+            renderedTag = [NSString stringWithFormat:@"%zd", pageNumber];
         }
         else
             renderedTag = [tag renderContentWithContext:context HTMLSafe:HTMLSafe error:error];
@@ -165,7 +165,7 @@ static NSArray * reportDefaultTags = nil;
     else if (sectionType == PRKSectionTypeFooter)
         footerFormatter = formatter;
     else
-        [NSException raise:@"Invalid Section Type" format:@"Section Type: %d is invalid", sectionType];
+        [NSException raise:@"Invalid Section Type" format:@"Section Type: %zd is invalid", sectionType];
 }
 
 - (void)renderPage
